Move Feature class into its own file
